server/requirements.txt
dashscope
flask
flask-cors
openai
python-dotenv
